chore(subgraphs/dex): refactor shared/numeric
