+++ subgraphs/dex/shared/libs/math.ts
@@ -1,0 +1,8 @@
+import { BigInt, BigDecimal } from "@graphprotocol/graph-ts";
+
+export const ZERO_BI = BigInt.fromI32(0);
+export const ONE_BI = BigInt.fromI32(1);
+export const ZERO_BD = BigDecimal.fromString("0");
+export const ONE_BD = BigDecimal.fromString("1");
+export const BI_18 = BigInt.fromI32(18);
+export const HALF_PERCENT = BigDecimal.fromString("0.5");
