--- subgraphs/dex/shared/numeric.ts
@@ -1,8 +1,0 @@
-import { BigInt, BigDecimal } from "@graphprotocol/graph-ts";
-
-export let ZERO_BI = BigInt.fromI32(0);
-export let ONE_BI = BigInt.fromI32(1);
-export let ZERO_BD = BigDecimal.fromString("0");
-export let ONE_BD = BigDecimal.fromString("1");
-export let BI_18 = BigInt.fromI32(18);
-export let HALF_PERCENT = BigDecimal.fromString("0.5");
