+++ subgraphs/dex/tests/.latest.json
@@ -1,4 +1,4 @@
 {
     "version": "0.5.3",
-    "timestamp": 1661619563803
+    "timestamp": 1661772563654
 }
